# chore: better performance on `array_append`