## Description

The `array_append` function was slow for building large arrays due to an inefficient deep-copy method that serialized the array to a string and back. An internal benchmark improved from >60s to <9s with this change.

This PR addresses the bottleneck and fixes related array-handling bugs:
- Replaced the string-based copy in `array_append` with `copyArray`, an efficient, type-aware memory copy.
- Improved the slice growth strategy in `setArr` to reduce reallocations.
- Fixed `arrayOfNulls` to correctly convert to a typed array when a non-null element is set or appended.

## Related Issue

- fix #1522

## Motivation and Context

The previous performance of `array_append` made building large arrays in Kuneiform actions impractical. This optimization is critical for real-world use cases that rely on array manipulation.

## How Has This Been Tested?

- Updated existing unit tests for array functions.
- Added new tests for `nil` and untyped array edge cases.
- Updated tests to reflect the new error propagation behavior.
- All new and existing tests pass.

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation update (including inline docs)
- [x] Other (please describe): Performance optimization.

## Checklist:
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.

### Checklist Explanation:
The changes are internal performance optimizations and bug fixes. No user-facing Kuneiform syntax or engine APIs were changed, so documentation updates are not required.

## How to Review this PR:
1.  Start with the new `copyArray` function in `node/engine/interpreter/values.go`.
2.  Review the updated `array_append` implementation in `node/engine/interpreter/interpreter.go`.
3.  Examine the changes to `setArr` and `arrayOfNulls` in `values.go`.
4.  Look at the updated tests in `interpreter_test.go` to see the new error propagation behavior and edge case handling.