@outerlook 2025-06-09
This PR was built only for testing purposes, as to really investigate and merge demands more effort to understand the impact.

but at least on the happy path, it works.

So I'll put this on hold until we can prioritize the performance of our operations again

But this has a significant impact on all of our operations (insert, reading data, etc)

cc @MicBun