@outerlook 2025-06-12
testing https://github.com/trufnetwork/node/blob/main/internal/benchmark/load_test.go

with 400 streams, 5 years of daily data (5 x 365 data points), `get_index_change`

before this PR:

```log
Qty Streams: 400, Branching Factor: 8, Data Points: 1825, Visibility: Public, Procedure: get_index_change, Samples: 1, Memory Usage: 1691 MB
  Mean Duration: 36.800188194s
  Min Duration: 36.800188194s
  Max Duration: 36.800188194s
```

after this PR

```log
Qty Streams: 400, Branching Factor: 8, Data Points: 1825, Visibility: Public, Procedure: get_index_change, Samples: 1, Memory Usage: 1691 MB
  Mean Duration: 9.097067288s
  Min Duration: 9.097067288s
  Max Duration: 9.097067288s
```


What other kinds of queries will share the improvements?
queries that have array_append used inside a for..loop for many loops